Updated Paragraph 1
I don't know what programming language you use, but my experience of recent coding in Java, Ruby, JavaScript, PHP, Python, C++, and Rust tells me that the principle, which I will try to convince you to adhere to --- is universal for all languages. It's about the naming of test files. It may look to you like a question of low importance, but let me try to demonstrate that it's not. How do you name your files with test classes? How many of them do you create in the src/test/java directory? Where do you place a class that is used only in a test but is not a test by itself? To most of these questions, the answer most of you would give is "Whatever!" So let's try to find a better answer.